refactor(api): consolidate user profile options and existence check interfaces into user API types
